nutri-health-api/requirements.txt
# FastAPI and server
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.11

# File handling
python-multipart==0.0.12
pillow==12.2.0

# AI Integration (OpenAI primary, Qwen-VL fallback via DashScope)
openai>=1.40.0,<2.0.0

# Environment variables
python-dotenv==1.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.12.5
pydantic-settings==2.13.1

# LangChain / RAG
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
langchain-huggingface==0.1.0
faiss-cpu==1.8.0; python_version < "3.13"
faiss-cpu==1.12.0; python_version >= "3.13"
tiktoken>=0.7.0
sentence-transformers>=2.6.0,<5.0.0
